global command arguments
Taking the great work the @tpett did for limiting memory usage
on commands that Pruview runs, this is just a name change to
better describe what is being configured.  This also means this
can be a place to add additional global command arguments as the
needs arises in the future.

To take this even further, it might be nice to build in a way for
users to configure these types of args more elegantly.  Right now,
the only way to change the default `GLOBAL_CMD_ARGS` would be to
duck-punch pruview.