update to a modern sRGB profile and cleanup applying the profile
This builds on the previous PR (always forcing sRGB colorspace)
and updates to a modern sRGB profile.  This also cleans up the
implementation of applying the profile by adding it to the format
cmd.  This is a more proper way of applying the colorspace and
results in fewer system cmds being run.